Shift platform specific code into arm_linux module
